Endpoints development

+++ code/jvm/main/kotlin/com/isel/leic/ps/ion_classcode/repository/UsersRepository.kt
@@ -10,7 +10,7 @@ interface UsersRepository {
     fun getUserById(id: Int): User?
     fun getUserByEmail(email: String): User?
     fun getUserByToken(token: String): User?
-    fun getUserByGithubId(githubId: Int): User?
+    fun getUserByGithubId(githubId: Long): User?
     fun deleteStudent(id: Int)
     fun deleteTeacher(id: Int)
 }
