+++ code/jvm/main/kotlin/com/isel/leic/ps/ion_classcode/repository/jdbi/JdbiUsersRepository.kt
@@ -91,7 +91,7 @@ class JdbiUsersRepository(
         return helper(handle = handle, id = id)
     }
 
-    override fun getUserByGithubId(githubId: Int): User? {
+    override fun getUserByGithubId(githubId: Long): User? {
         val id = handle.createQuery(
             """
             SELECT id FROM users
